refactor(api): migrate errors used by migrated services file
