+++ api/src/access/authentication/domain/errors.js
@@ -1,0 +1,22 @@
+import { DomainError } from '../../../shared/domain/errors.js';
+
+class MissingOrInvalidCredentialsError extends DomainError {
+  constructor(message = 'Missing or invalid credentials') {
+    super(message);
+  }
+}
+
+class PasswordNotMatching extends DomainError {
+  constructor(message = 'Wrong password.') {
+    super(message);
+  }
+}
+
+class UserShouldChangePasswordError extends DomainError {
+  constructor(message = 'User password must be changed.', meta) {
+    super(message);
+    this.meta = meta;
+  }
+}
+
+export { MissingOrInvalidCredentialsError, PasswordNotMatching, UserShouldChangePasswordError };
