+++ api/src/access/authentication/domain/usecases/authenticate-user.js
@@ -4,7 +4,10 @@ const { get } = lodash;
 
 import { PIX_ORGA, PIX_ADMIN } from '../../../../../lib/domain/constants.js';
 import { ForbiddenAccess, LocaleFormatError, LocaleNotSupportedError } from '../../../../shared/domain/errors.js';
-import { MissingOrInvalidCredentialsError, UserShouldChangePasswordError } from '../../../shared/domain/errors.js';
+import {
+  MissingOrInvalidCredentialsError,
+  UserShouldChangePasswordError,
+} from '../../../authentication/domain/errors.js';
 
 async function _checkUserAccessScope(scope, user, adminMemberRepository) {
   if (scope === PIX_ORGA.SCOPE && !user.isLinkedToOrganizations()) {
