--- api/src/access/shared/domain/errors.js
@@ -1,33 +1,0 @@
-import { DomainError } from '../../../shared/domain/errors.js';
-
-class ApplicationWithInvalidClientSecretError extends DomainError {
-  constructor(message = 'The client secret is invalid.') {
-    super(message);
-  }
-}
-
-class MissingOrInvalidCredentialsError extends DomainError {
-  constructor(message = 'Missing or invalid credentials') {
-    super(message);
-  }
-}
-
-class PasswordNotMatching extends DomainError {
-  constructor(message = 'Mauvais mot de passe.') {
-    super(message);
-  }
-}
-
-class UserShouldChangePasswordError extends DomainError {
-  constructor(message = 'Erreur, vous devez changer votre mot de passe.', meta) {
-    super(message);
-    this.meta = meta;
-  }
-}
-
-export {
-  ApplicationWithInvalidClientSecretError,
-  MissingOrInvalidCredentialsError,
-  PasswordNotMatching,
-  UserShouldChangePasswordError,
-};
