+++ api/tests/integration/application/error-manager_test.js
@@ -1,10 +1,16 @@
 import { expect, HttpTestServer, sinon } from '../../test-helper.js';
 import * as DomainErrors from '../../../lib/domain/errors.js';
-import { CsvImportError, EntityValidationError, ForbiddenAccess } from '../../../src/shared/domain/errors.js';
+import {
+  CsvImportError,
+  EntityValidationError,
+  ForbiddenAccess,
+  InvalidResultRecipientTokenError,
+  InvalidTemporaryKeyError,
+} from '../../../src/shared/domain/errors.js';
 import {
   MissingOrInvalidCredentialsError,
   UserShouldChangePasswordError,
-} from '../../../src/access/shared/domain/errors.js';
+} from '../../../src/access/authentication/domain/errors.js';
 
 describe('Integration | API | Controller Error', function () {
   let server;
@@ -818,7 +824,7 @@ describe('Integration | API | Controller Error', function () {
     });
 
     it('responds Unauthorized when a InvalidTemporaryKeyError error occurs', async function () {
-      routeHandler.throws(new DomainErrors.InvalidTemporaryKeyError('Demande de réinitialisation invalide.'));
+      routeHandler.throws(new InvalidTemporaryKeyError('Demande de réinitialisation invalide.'));
       const response = await server.requestObject(request);
 
       expect(response.statusCode).to.equal(UNAUTHORIZED_ERROR);
@@ -826,7 +832,7 @@ describe('Integration | API | Controller Error', function () {
     });
 
     it('responds Unauthorized when a InvalidResultRecipientTokenError error occurs', async function () {
-      routeHandler.throws(new DomainErrors.InvalidResultRecipientTokenError());
+      routeHandler.throws(new InvalidResultRecipientTokenError());
       const response = await server.requestObject(request);
 
       expect(response.statusCode).to.equal(UNAUTHORIZED_ERROR);
