+++ api/tests/integration/application/passwords/password-controller_test.js
@@ -1,14 +1,10 @@
 import { domainBuilder, expect, sinon, HttpTestServer } from '../../../test-helper.js';
 
-import {
-  InvalidTemporaryKeyError,
-  PasswordResetDemandNotFoundError,
-  UserNotFoundError,
-} from '../../../../lib/domain/errors.js';
+import { PasswordResetDemandNotFoundError, UserNotFoundError } from '../../../../lib/domain/errors.js';
 
 import { usecases } from '../../../../lib/domain/usecases/index.js';
 import * as moduleUnderTest from '../../../../lib/application/passwords/index.js';
-import { ForbiddenAccess } from '../../../../src/shared/domain/errors.js';
+import { ForbiddenAccess, InvalidTemporaryKeyError } from '../../../../src/shared/domain/errors.js';
 
 describe('Integration | Application | Passwords | password-controller', function () {
   let httpTestServer;
