+++ api/tests/shared/integration/infrastructure/repositories/admin-member-repository_test.js
@@ -2,7 +2,8 @@ import { expect, databaseBuilder, knex, catchErr, sinon } from '../../../../test
 import { PIX_ADMIN } from '../../../../../lib/domain/constants.js';
 import * as adminMemberRepository from '../../../../../src/shared/infrastructure/repositories/admin-member-repository.js';
 import { AdminMember } from '../../../../../lib/domain/models/AdminMember.js';
-import { AdminMemberError } from '../../../../../lib/domain/errors.js';
+
+import { AdminMemberError } from '../../../../../src/access/authorization/domain/errors.js';
 
 const { ROLES } = PIX_ADMIN;
