+++ api/tests/unit/domain/errors_test.js
@@ -5,13 +5,15 @@ import {
   CsvImportError,
   EntityValidationError,
   ForbiddenAccess,
+  InvalidTemporaryKeyError,
   LocaleFormatError,
   LocaleNotSupportedError,
 } from '../../../src/shared/domain/errors.js';
+import { AdminMemberError } from '../../../src/access/authorization/domain/errors.js';
 
 describe('Unit | Domain | Errors', function () {
   it('should export a AdminMemberError', function () {
-    expect(errors.AdminMemberError).to.exist;
+    expect(AdminMemberError).to.exist;
   });
 
   it('should export a CertificationCandidateAlreadyLinkedToUserError', function () {
@@ -160,7 +162,7 @@ describe('Unit | Domain | Errors', function () {
 
   describe('#InvalidTemporaryKeyError', function () {
     it('should export a InvalidTemporaryKeyError', function () {
-      expect(errors.InvalidTemporaryKeyError).to.exist;
+      expect(InvalidTemporaryKeyError).to.exist;
     });
 
     it('should have a getErrorMessage method', function () {
@@ -172,7 +174,7 @@ describe('Unit | Domain | Errors', function () {
       };
 
       // then
-      const error = new errors.InvalidTemporaryKeyError();
+      const error = new InvalidTemporaryKeyError();
       expect(error.getErrorMessage).to.be.a('function');
       expect(error.getErrorMessage()).to.eql(expectedErrorMessage);
     });
