+++ api/tests/unit/domain/usecases/get-user-by-reset-password-demand_test.js
@@ -1,13 +1,10 @@
 import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
 import { User } from '../../../../lib/domain/models/User.js';
 
-import {
-  InvalidTemporaryKeyError,
-  PasswordResetDemandNotFoundError,
-  UserNotFoundError,
-} from '../../../../lib/domain/errors.js';
+import { PasswordResetDemandNotFoundError, UserNotFoundError } from '../../../../lib/domain/errors.js';
 
 import { getUserByResetPasswordDemand } from '../../../../lib/domain/usecases/get-user-by-reset-password-demand.js';
+import { InvalidTemporaryKeyError } from '../../../../src/shared/domain/errors.js';
 
 describe('Unit | UseCase | get-user-by-reset-password-demand', function () {
   const temporaryKey = 'ABCDEF123';
